remove unused code and simplify
